Conversion to JSON and more
I finally finished the conversion from pickle to json. I changed the animelists to dictionaries with the index and anime name in it to make sure the program doesn't cry when there are anime with the same name. The add function is a lot shorter and useful. I changed the load so it doesnt save the animelist every time, but instead it saves the file whenever you add an anime or remove one. Added a save function for simplicity's sake. The code has already become more readable but there is a lot of work that remains. The last thing i added was an options dictionary. mostly for later.