Added settings
I added the settings function so you can change the settings and it will automatically reload the application afterwards. Cleaned up some imports, will do more later. Everything needed for the startup is now in the start function. Made a save setting function, a function to clarify which list you'd like to use, a load list option, begin2 is now gone, trueorfalsesetting was made for all Boolean settings, listsettings was made for the settings that require you to use lists so it makes it impossible to choose a non existing animelist, which setting was made to clarify which setting you want to change (obviously), and last but not least setting was added to change the settings.