Update from yesterday
I've finished the last things from my todolist from yesterday. shortened the page soup getter so it gives an error if the page doesn't exist. The changes function is now shorter since i made a few extra functions to make finding stuff on a webpage easier. I also made a function to look for the missing animes that you ever looked for. This is just so it's possible to sort by the new information. If an anime is alread in another list it will ask if you want to change the anime to the list you are currently on. The delete function has been fixed. The sort function was a tough fix, but it finally works, i'm probably overthinking it, but i always do that so that's fine, ill look at it later. Change list name function has been fixed and you can now change the sorting method.