Change Episodes
You can now change the amount of episodes you've watched. the program now sorts the list every time it changes or when you change the sorting method. I made a to do list for the sorting methods that don't work yet. The checklistmaker is not list bound anymore as well, so it's functional for the animelist too. I also added the anime from the list Chewie to the digital version.