added functions + list chewie
I added a function that shows the total watch time, and a function that allows you to add anime from folder names. Don't think it would be useful to anyone except for me since I have a folder of 226 folders with anime songs in them. At last I digitalized the chewie list.